pyproject.toml
[project]
name = "copilot-orchestrator"
version = "0.6.0"
description = "Orchestration Layer for the Copilot system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.0.10",
    "openai>=1.0.0",
    "fastapi[standard]>=0.110.0",
    "uvicorn>=0.23.0",
    "rich>=13.0.0",
    "langfuse[langchain]>=2.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "redis>=5.0.0",
    "langgraph-checkpoint-redis>=0.1.0",
    "langchain>=1.2.15",
    "mcp>=1.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
    "commitizen>=3.0.0",
    "types-python-jose",
    "types-redis"
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "C4",  # flake8-comprehensions
    "PL",  # pylint
    "RUF"  # ruff-specific rules
]
ignore = ["PLR0913"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
exclude = ["tests", "venv", ".venv"]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.6.0"
tag_format = "v$version"
update_changelog_on_bump = false
version_files = [
    "pyproject.toml:version"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"

[tool.uv]
managed = true
